requirements.txt
python-telegram-bot==21.3
binance-futures-connector==4.0.0
pandas==2.1.4
numpy==1.26.4
ta==0.11.0
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.5
